requirements.txt
fastapi==0.111.0
pydantic==2.7.1
uvicorn==0.29.0
PyYAML==6.0.1
PyMySQL==1.1.1
openai==1.30.3
requests==2.32.3
boto3==1.34
google-cloud-speech==2.26.0
pydub==0.25.1
psycopg2-binary